Implement generic job to generate annotation patches
